requirements.txt
paramiko==3.1.0
PyYAML==6.0
